refac: UserTable 분리
