+++ test-web/src/components/UserManagement/Childeren/handleUserEdit.js
@@ -9,14 +9,13 @@ const handleUserEdit = async (params, handleSnackbarShow) => {
     userId: id,
     [field]: value,
   };
-
   try {
     // Send a POST request to update the user's information
     const response = await userUpdate(req);
     if (response.ok) {
       const data = await response.json();
       handleSnackbarShow(
-        `${data.name}님의 ${field}이(가) ${value}로 수정되었습니다.`,
+        `${data.name}님의 ${field}이 ${value}로 수정되었습니다.`,
         'success'
       );
       return true;
